made directories to separate different kinds of tests
