+++ projectApp/tests/logConfirmationTests/2DGraph.test.ts
@@ -7,7 +7,7 @@ import { fileContainsText } from "./fileParser";
 describe('2D graph log tests', () => {
     // The log in question is located here
     // This log is used to avoid build issues
-    const filePath = "./logConfirmationlogs.txt"
+    const filePath = "./tests/logConfirmationTests/logConfirmationlogs.txt"
 
     // Check if the addPoint() log calls are being reached
     test('file contains addPoint() logging', async() => {
