+++ projectApp/tests/logConfirmationTests/logConfirmationlogs.txt
