+++ projectApp/tests/unit_tests/csvTesting.test.ts
@@ -1,8 +1,8 @@
 import { describe } from 'node:test';
 import { expect, test } from 'vitest';
 
-import { LocalCSVReader as localReader, LocalCSVHeaders as localHeaders, UrlCSVReader as urlReader, UrlCSVHeaders as urlHeaders } from "../src/components/Csv_Components/CSVReaders.tsx";
-import { CSVHeaders, TimeSeriesData } from '../src/types/CSVInterfaces.tsx';
+import { LocalCSVReader as localReader, LocalCSVHeaders as localHeaders, UrlCSVReader as urlReader, UrlCSVHeaders as urlHeaders } from "../../src/components/Csv_Components/CSVReaders.tsx";
+import { CSVHeaders, TimeSeriesData } from '../../src/types/CSVInterfaces.tsx';
 
 
 interface ReaderTest<T> {
